Ignore dangling messages from previous runs
This commit fixes cases where a dangling message would cause a
subsequent run to fail.

This would happen when a timeout occurs, but the messages still get
received by the calling process. Then on the next call, `SSHEx` would
not differentiate between the messages of the previous run and the
messages of the current run.

This commit fixes it by making the messages unambiguous, making sure
that they belong to the `conn` and `channel` being used on the current
call.

It also removes the compatibility with the `{:error, reason}` message
format, as it is not a message the `:ssh` module sends, and it severely
hinders the ability to differentiate between messages from the current
and previous runs.